pyproject.toml
[project]
name = "chr2025-seeing-history-unseen"
version = "0.1.0"
description = "CHR 2025 conference materials: Seeing History Unseen - Evaluating Vision-Language Models for WCAG-Compliant Alt-Text in Digital Heritage Collections"
authors = [
    { name = "Moritz Mähr", email = "" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[project.scripts]
benchmark = "src.benchmarking.cli:cli"
